Add logging to NALD import licence jobs (#677)
https://eaflood.atlassian.net/browse/WATER-4024
https://eaflood.atlassian.net/browse/WATER-4039

As part of our efforts to [clean up the logs](https://github.com/DEFRA/water-abstraction-import/pull/654) for the import service we removed the logging that was happening in `src/modules/nald-import/jobs/import-licence.js`.

One job gets added for every licence to import which is normally 71K! Since then we've been looking at the output to confirm we can see what is going on.

We found there was a big unknown when it came to the NALD import licence job. With the logging removed, there was no way to confirm the jobs were running. We didn't want to put the logging back in. But also, we were limited with what we could do. Later versions of PGBoss allow you to query the queue and jobs remaining but our version is too old and out of scope for being updated.

So, the solution we have come up with is to add data on how many invoices there are to import, along with the index position of each one to the job's data. We can then use this information to decide whether to output a start and finish log message that confirms to us this process is happening.

For reference, the import can take _up to 3 hours!_ So, when you are checking the logs be patient waiting for the 'finished' entry 😬