Remove assert tables exist in NALD import job
https://eaflood.atlassian.net/browse/WATER-4024
https://eaflood.atlassian.net/browse/WATER-4039

When looking at the [logging in](https://github.com/DEFRA/water-abstraction-import/pull/677) `src/modules/nald-import/jobs/import-licence.js` we spotted that it runs a check to confirm the import tables exist.

The issue is this job isn't a one-off. We create a job for every licence to be imported which is normally 71K of them! That means we're doing the same check 71 thousand times. The current process takes approximately 3 to 4 hours to complete. So, every saving we can make will help.

It doesn't help that the `src/modules/nald-import/jobs/populate-pending-import.js` which is responsible for determining which licences will be imported already does the check 🤦.

So, we're dropping it from `src/modules/nald-import/jobs/import-licence.js` to see if we can speed things up a little bit.