Tidy up inconsistences in Licence & NALD imports
Whilst working on [Consolidate NALD Import jobs](https://github.com/DEFRA/water-abstraction-import/pull/679) and [Consolidate Licence Import jobs](https://github.com/DEFRA/water-abstraction-import/pull/682) we noticed some inconsistencies between them. Also, as we get to understand what they are doing some of the naming didn't make sense.

As the 2 primary imports in the service we took the time to iron out some of the inconsistencies and try and improve the naming of some of the jobs.