Fix Airbrake when behind a proxy
https://github.com/DEFRA/water-abstraction-team/issues/104

In the change [Replace legacy logging](https://github.com/DEFRA/water-abstraction-import/pull/654) we removed the deprecated [airbrake-js](https://www.npmjs.com/package/airbrake-js) implementation and updated it to use the latest [@airbrake/node](https://github.com/airbrake/airbrake-js/tree/master/packages/node).

We copied the plugin we have in [water-abstraction-system](https://github.com/DEFRA/water-abstraction-system). It works great locally and we just assumed both **water-abstraction-system** and **water-abstraction-import** weren't firing any errors in production. But we've realised no errors are getting through in our AWS environments because the apps sit behind a proxy and we haven't configured Airbrake to handle this.

In this change we follow the Airbrake docs and properly configure Airbrake (and yes - we have tested it this time!)