Replicate NALD returns when import triggered
https://eaflood.atlassian.net/browse/WATER-4270

We've been working on building a better understanding of the jobs within the import service. We've even documented [details on the background jobs](https://github.com/DEFRA/water-abstraction-team/pull/39).

One of the things we'd like to be able to do is rebuild a DB locally from scratch. This will be especially useful for those new to the team when setting up their environment (and the rest of us when we break it!)

Something we noticed when trialling how to do this was that there was no return submission data. This would be really useful to have and should be available in NALD for importing. Well, it turns out it is but it's hidden behind a hard-coded 'config' value (🤦🤡)

When we manually changed the value and ran the job our locally DB was suddenly populated with lots of return information which makes testing much easier and the local experience more realistic.

Reviewing the commit history and the existing comment (`// Set to false as this is highly disruptive`) we understand this is not something you want to happen every time the job is run (which is at least once a day). But we can make a fair assumption it is something we do want to happen when we manually trigger the NALD import. Our only reason for doing so is when rebuilding the DB.

This change makes the necessary amendments to allow us to automatically replicate returns data when manually triggering the NALD import.