First step: pass a value to the first job
Unfortunately, this being done via 'jobs' we've got to get our value from the initial POST request to the last job in the chain. This means we need to pass the param between jobs.

This is the first step, ensuring the `replicateReturns` param is included in the s3-download job's data.