Final step: pass the value to where it is needed
Did you think once we got to the final job we'd be done!?

The final step is about getting the value to where the logic currently refers to the hard-coded config value. That unfortunately means passing it through a number of modules and functions. But we got it there eventually!