Fix error logging and notifications
https://eaflood.atlassian.net/browse/WATER-4039

In [Replace legacy logging](https://github.com/DEFRA/water-abstraction-import/pull/654) we switched from the way the app was previously logging to how we do things in [water-abstraction-system](https://github.com/DEFRA/water-abstraction-system).

Or so we thought! If an error happens we've been getting a log message and a notification to Errbit. However, we haven't been getting the correct stack trace.

Digging in we've spotted it is because the order of args in the base notifier do not line up with how the rest of the code is calling `.omfg()`. Doh! 🤦

So, the base notifier has been wrapping the error in another error hence muddle stack traces! This change things the base notifier to line up with the calls so we can start getting clearer, more informative stack traces again!