Fix error logging and notifications (#919)
https://eaflood.atlassian.net/browse/WATER-4039

In [Replace old logging](https://github.com/DEFRA/water-abstraction-import/pull/654), we migrated from the app's prior logging method to the [water-abstraction-system](https://github.com/DEFRA/water-abstraction-system) approach.

Or so we thought! If an error occurs, we receive a log message and a notification to Errbit. However, we have not been receiving the correct stack trace.

Digging deeper, we discovered that the order of args in the base notifier does not match how the remainder of the code calls '.omfg()'. Doh! 🤦

So, the base notifier has been encapsulating the error in another error, causing muddled stack traces! This changes the base notifier to align with the calls, allowing us to receive cleaner, more informative stack traces again!