Add collection_frequency to import return requirements query (#925)
https://eaflood.atlassian.net/browse/WATER-4473

As part of the work to get ready to replace NALD for handling returns we need to be able to collect the collection frequency for a return. This PR adds the collection frequency to the return requirements import job. We also noticed the the reporting frequency was using the wrong variable so we corrected that as well.