Add notes to return-versions from return-requirements (#929)
https://eaflood.atlassian.net/browse/WATER-4472

As part of our work to get ready to replace NALD for handling returns we need to be able to let users store a note against the return. This is currently handled at the return-requirement description but we want to do it in the return-version level. This PR updates the import job to concatenate the descriptions from the return-requirements table and put them in the notes section of the return-version