Include NALD point ID in return reqs import (#942)
https://eaflood.atlassian.net/browse/WATER-4493

Returns submitted to the service are sent back to NALD. There are then data reports that link the return submissions to their relevant points. But when we take over the returns functionality from NALD the link between a return submission (form log) and the point will be lost if we don't have some way of capturing the NALD unique point ID.

The submissions (returns.returns) link to `water.return_requirements` which in turn is linked to `water.return_requirement_points`. If that table contains the NALD point ID then they should be able to link submissions to specific points when creating reports.

In [Add NALD point ID to return req points table](https://github.com/DEFRA/water-abstraction-service/pull/2552) we made sure the table has a new field to hold the point ID. In this change, we alter the import job to ensure it gets populated.