Fix create notes from RTN FORMATS desc for versions
https://eaflood.atlassian.net/browse/WATER-4472

As part of migrating the return requirement functionality from NALD to WRLS we've had to expand how the existing legacy code is importing the current from NALD.

One of the changes we made was to capture the descriptions from each return version's return formats (return requirements) and combine them as the note. We don't intend to use `description` at the return requirement level but this is a way of still being able to capture the info and make them visible to users should they need them.

However, we hadn't catered for the fact the files we're importing from state `'null'` for empty fields. This means `'null'` gets imported rather than the field being left as `NULL`. This messed up the notes because where there was no return format description `'null'` was being used, resulting in notes like 'null, null'.