Transform return version status to WRLS style
https://eaflood.atlassian.net/browse/WATER-4505

We've been extended and amended the import of return versions from NALD to WRLS as part of our work to switch from NALD to WRLS for managing them.

One of the changes is what the status is expected to be. In NALD as soon as a new return version is added the previous one is marked as superseded. 'current' means "this is the latest version". In WRLS if a new version starts later than the previous version it leaves the previous record alone. 'current' means "all versions with a status of current are still valid". If a new version starts on the same date as the previous one, then the new version 'supersedes' the previous one so we update the status of the previous one.

To support this we need to make a change to the import. Once the records have been updated we'll run a new query which will determine those that are truly superseded from those that are not. It will update the status of those that are _not_ truly 'superseded' to 'current'.