Set to 'draft' ret. versions with no requirements (#950)
https://eaflood.atlassian.net/browse/WATER-4510

We've been extending and amending the import of return versions from NALD to WRLS as part of our work to switch from NALD to WRLS for managing them.

In WRLS, we want to allow users to add new return requirements by copying them from a previous 'current' return version. We also want to display and view them in the UI.

However, we know our new functionality will have to support working with the existing data that was created and vetted in NALD. So, we double checked, and though only a few, we found instances of WRLS 'current' return versions that have no return requirements.

These are not valid records; we cannot copy from them, display them in the UI, or generate return logs. We also don't want to have to consider them when new return versions are added and we have to update the previous ones.

So, this change adds a new query to the return versions import in charging data to update the status to 'draft' those return versions with a status of 'current' that have no return requirements.

It caters for the fact that there will be times it _is_ valid to have a return version that is 'current' but has no return requirements. These will be how we reflect return versions created for licences where no-returns are required. This is a feature that NALD doesn't have, but WRLS will. We expect that the import will be switched off before the new functionality is switched on. But just in case, the query ignores those return versions created with a no-returns required `reason`.