Add missing return version end dates during import
https://eaflood.atlassian.net/browse/WATER-4514

We've been extending and amending the import of return versions from NALD to WRLS as part of our work to switch from NALD to WRLS for managing them.

We've spotted that some of the return versions we are importing are missing end dates.

|Return Version ID|Licence ID|Version number|Start date|End date  |Status |
|-----------------|----------|--------------|----------|----------|-------|
|RV1              |LV1       |100           |1984-10-22|2008-03-31|Current|
|RV2              |LV1       |101           |2008-04-01|2009-03-31|Current|
|RV3              |LV1       |102           |2009-04-01|          |Current|
|RV4              |LV1       |103           |2023-04-01|          |Current|

We expect the 'latest' return version to have no end date because it reflects what is the return requirements that should be used for anything in the future. But in this example `RV3` should have an end date. This is an error on the part of NALD we should correct during the import.