Add reason mapping query to return version import
https://eaflood.atlassian.net/browse/WATER-4636

> Part of the work to display a licence's history to users (mod log)

Now that we are [importing the NALD mod logs](https://github.com/DEFRA/water-abstraction-import/pull/998) and linking them to their WRLS version records (charge, licence and returns) we can look to update the WRLS records with their missing reasons.

We've done some work with the Billing & Data team to map NALD reasons to those in WRLS. There are a couple that don't map, and for those [water-abstraction-system](https://github.com/DEFRA/water-abstraction-system) will fall back to whatever reason description we have in the mod log. For the rest though, we can fill in their missing `reason` field with the mapped WRLS reason. We can then start displaying this to users in the out new UI screens.

This change adds another query to the charging data import, that will do the mapping for those WRLS return versions that

- have an empty `reason``
- have mod log records
- the first mod log record has a reason code
- the reason code is mapped to WRLS