Fix return version note creation from requirements
https://eaflood.atlassian.net/browse/WATER-4472

> Part of the work to migrate return version management from NALD to WRLS

In NALD notes can be saved at the return requirement level. But in WRLS they will be held at the return version level.

- `return_version`
  - `return_requirement`
  - `return_requirement`

Previously, we made a change to the import to [Add notes to return-versions from return-requirements](https://github.com/DEFRA/water-abstraction-import/pull/929) by concatenating what we find at the return requirement level as the return version note.

But we've spotted the query we wrote was flawed. It is not taking into account the region code. NALD, for whatever reason, doesn't generate a unique ID for each record. This means you will see the following in its `NALD_RET_VERSIONS` table.

| id   | region |
|------|--------|
|100101|      6 |
|100101|      2 |
|100101|      3 |

You have to combine the 2 values to make a unique ID. The same goes for `NALD_RET_FORMATS` (return requirements).

Our problem is the query we wrote selected the return requirements just using the ID. This means notes from return requirements linked to other return versions are getting mixed into the notes we're generating.

This change corrects the query. As it is run during every import all records will be fixed when next run.