Fix return version note creation from requirements (#1004)
https://eaflood.atlassian.net/browse/WATER-4472

> Part of the work to migrate return version management from NALD to WRLS

Notes can be saved at the return requirement level in NALD, but they will be held at the return version level in WRLS.

Previously, we changed the import to [Add notes to return-versions from return-requirements](https://github.com/DEFRA/water-abstraction-import/pull/929) by concatenating what we find at the return requirement level as the return version note.

- `return_version` - _Note 1, Note 2_
  - `return_requirement` - _Note 1_
  - `return_requirement` - _Note 2_

But we've spotted that the query we wrote was flawed. It does not take into account the region code. NALD, for whatever reason, doesn't generate a unique ID for each record. This means you will see the following in its `NALD_RET_VERSIONS` table.

| id   | region |
|------|--------|
|100101|      6 |
|100101|      2 |
|100101|      3 |

You have to combine the 2 values to make a unique ID. The same goes for `NALD_RET_FORMATS` (return requirements).

Our problem is the query we wrote selected the return requirements just using the ID. This means notes from return requirements linked to other return versions are getting mixed into the notes we're generating.

This change corrects the query. It is run during every import, so all records will be fixed during the next run.