Add new import points job to licence import
When we amended the import to bring in return requirement points we were able to tag on another query to those already being run as part of the 'charging-data' job.

Unfortunately, for 'reasons', the previous team made the licence import job much more complicated. It is broken down into various 'jobs', which typically

- queue up a bunch of jobs, for example, licences to import
- process each individual licence import job

We're in the middle of migrating the licence import to [water-abstraction-system](https://github.com/DEFRA/water-abstraction-system) which will do away with all this needless complexity. But that won't 'land' for some time and we need points now.

This is a half-way house. Adding this licence-based query to the 'charging-data' job didn't seem right. But replicating how things are done in the 'licence-import' job seemed overkill.

So, we've amended the job process in licence import to queue up a new 'import-points' job when the last licence has been imported. But the job is just running a single query to create or update the points in one go.