Remove note reset in Ret. Vers. note update
> Part of the work to migrate return version management from NALD to WRLS

In [Fix return version note creation from requirements](https://github.com/DEFRA/water-abstraction-import/pull/1004) we amended the way we were updating `water.return_versions` notes from `water.return_requirements` descriptions having spotted a flaw in the original logic.

But because the flaw shipped, we needed to reset all notes before running the new version of the query. We only needed to run the reset once, but we don't have the luxury of doing that easily in production!

So, instead we had the notes update query always run the reset before then updating the notes.

The fix has [now been released](https://github.com/DEFRA/water-abstraction-import/releases/tag/v2.30.0) and we've confirmed it corrected the return version notes. So, we can remove the reset from the query we are running.