Remove note reset in Ret. Vers. note update (#1013)
https://eaflood.atlassian.net/browse/WATER-4472

> Part of the work to migrate return version management from NALD to WRLS

In [Fix return version note creation from requirements](https://github.com/DEFRA/water-abstraction-import/pull/1004), we amended the way we updated `water.return_versions` notes from `water.return_requirements` descriptions after spotting a flaw in the original logic.

However, because the flaw shipped, we needed to reset all notes before running the new version of the query. We only needed to run the reset once, but we don't have the luxury of doing that easily in production!

So, instead, we had the notes update query always run the reset before updating the notes.

The fix has [now been released](https://github.com/DEFRA/water-abstraction-import/releases/tag/v2.30.0), and we've confirmed it corrected the return version notes. So we can remove the reset from the query we are running.