Update Ret. vers. end-dates query to handle new
https://eaflood.atlassian.net/browse/WATER-4514

> Part of the work to migrate return version management from NALD to WRLS

In [Add missing return version end dates during import](https://github.com/DEFRA/water-abstraction-import/pull/954) we added a query to the import that will fill in the missing end dates for return versions imported from NALD. WRLS requires a 'complete' history in the versions (no gaps) which isn't enforced in NALD.

It works great! The problem is it didn't consider returns versions added directly to WRLS. In fairness, it didn't need to. The intent is that the import would be killed on the day WRLS took over return version management.

In practice though, the team and our users are trying to conduct testing of the new functionality, but then seeing what they've added messed with when the import runs during the night.

This change updates the query to ignore stuff added directly in WRLS, and just look at the records imported from NALD to stop things getting messed up.