Stop updating our corrected records!
Realised in testing these changes that there is another source of data getting messed up.

Where we have not added any new return versions to a licence the following has been happening;

- import runs
- finds existing records so updates their dates and status because of the ON CONFLICT()
- our clean up scripts run and repopulate missing end dates

We expect the last imported return version to not have an end date and our clean up queries are not filling them in. So, the latest return versions would have a blank end date.

Then, a user adds a new return version via WRLS. When that is added it will populate the end date of the last return version imported to be start date minus one day.

A bit later, it is night and the import runs again

- import runs
- finds existing records so updates their dates and status because of the ON CONFLICT()
- argh! It has reset the end date to NULL
- our clean up scripts run but don't consider this record as having a missing end date

Now the user goes in and finds the second record no longer has an end date. Plus our downstream functions start failing because they can no longer determine which return version covers a particular period.